backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
python-jose[cryptography]
passlib
sqlalchemy
python-multipart
bcrypt
